+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+7.96
+-----
+
+
 7.95
 -----
 - Fix the trending notification message [#3351](https://github.com/Automattic/pocket-casts-ios/pull/3351)